requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
pmdarima
scikit-learn
streamlit
python-dotenv
langchain
langchain-groq
groq
tqdm
plotly
